requirements.txt
fastapi[all]~=0.92.0
loguru~=0.6.0
psycopg[binary]
bcrypt~=4.0.1